refactor: move Farmhand component to components directory
