+++ src/components/ContextPane/ContextPane.js
@@ -1,7 +1,7 @@
 import React, { memo } from 'react'
 import { array, string } from 'prop-types'
 
-import FarmhandContext from '../../Farmhand.context'
+import FarmhandContext from '../Farmhand/Farmhand.context'
 import Inventory from '../Inventory'
 import CowPenContextMenu from '../CowPenContextMenu'
 import { stageFocusType } from '../../enums'
