+++ src/components/Farmhand/Farmhand.sass
@@ -1,6 +1,6 @@
 @import '~compass-mixins/lib/compass/css3/transform'
-@import styles/variables.sass
-@import styles/utils.sass
+@import ../../styles/variables.sass
+@import ../../styles/utils.sass
 
 // Needed to prevent pull-to-refresh on Android
 // https://stackoverflow.com/a/47709903/470685
@@ -101,7 +101,7 @@ body
     display: flex
 
   .sidebar
-    background-image: url('./img/ui/blue-stripe-bg.png')
+    background-image: url('../../img/ui/blue-stripe-bg.png')
     box-sizing: border-box
     overflow: auto
     padding: 1em 1em 0
@@ -283,7 +283,7 @@ body
       text-align: center
 
   .MuiDialog-paper
-    background-image: url('./img/ui/light-blue-stripe-bg.png')
+    background-image: url('../../img/ui/light-blue-stripe-bg.png')
 
   .MuiAlert-root
     margin-bottom: 1em
