+++ src/components/Toolbelt/Toolbelt.js
@@ -10,7 +10,7 @@ import { toolLevel } from '../../enums'
 
 import { memoize } from '../../utils'
 
-import FarmhandContext from '../../Farmhand.context'
+import FarmhandContext from '../Farmhand/Farmhand.context'
 import toolsData from '../../data/tools'
 
 import { tools as toolImages, craftedItems, pixel } from '../../img'
